Cree los repos, services y controllers para cancellations, cash movements, cash register, terminals y providers

+++ src/models/cancellation.model.js
@@ -11,7 +11,7 @@ const Cancellation = sequelize.define("Cancellation", {
 		allowNull: false,
 	},
 	amount: {
-		type: DataTypes.DECIMAL,
+		type: DataTypes.FLOAT,
 		allowNull: false,
 	},
 })
